[3591] Extract shared teacher search logic
