+++ spec/services/admin/teachers/search_spec.rb
@@ -8,13 +8,6 @@ RSpec.describe Admin::Teachers::Search do
   describe "#teacher_scope" do
     subject(:teacher_scope) { search.teacher_scope }
 
-    context "when the query is blank" do
-      let!(:teacher) { FactoryBot.create(:teacher) }
-      let!(:other_teacher) { FactoryBot.create(:teacher) }
-
-      it { is_expected.to contain_exactly(teacher, other_teacher) }
-    end
-
     context "when teachers have different creation times" do
       let!(:older_teacher) { FactoryBot.create(:teacher, created_at: 2.days.ago) }
       let!(:newer_teacher) { FactoryBot.create(:teacher, created_at: 1.day.ago) }
@@ -24,116 +17,12 @@ RSpec.describe Admin::Teachers::Search do
       end
     end
 
-    context "when it is an exact 7 digit TRN" do
-      let(:query_string) { "1234567" }
-      let!(:teacher) { FactoryBot.create(:teacher, trn: "1234567") }
-      let!(:other_teacher) { FactoryBot.create(:teacher, trn: "7654321", trs_first_name: "1234567", trs_last_name: "Teacher") }
-
-      it { is_expected.to contain_exactly(teacher) }
-    end
-
-    context "when it contains a full TRN with extra text" do
-      let(:query_string) { "TRN 1234567" }
-      let!(:teacher) { FactoryBot.create(:teacher, trn: "1234567") }
-      let!(:other_teacher) { FactoryBot.create(:teacher, trs_first_name: "TRN", trs_last_name: "Teacher") }
-
-      it { is_expected.to contain_exactly(teacher) }
-    end
-
-    context "when it is an exact API participant ID" do
-      let(:query_string) { "123e4567-e89b-12d3-a456-426614174000" }
-      let!(:teacher) { FactoryBot.create(:teacher, api_id: "123e4567-e89b-12d3-a456-426614174000") }
-      let!(:other_teacher) { FactoryBot.create(:teacher, api_id: "999e4567-e89b-12d3-a456-426614174999") }
-
-      it { is_expected.to contain_exactly(teacher) }
-    end
-
-    context "when it contains a full API participant ID with extra text" do
-      let(:query_string) { "API ID 123e4567-e89b-12d3-a456-426614174000" }
-      let!(:teacher) { FactoryBot.create(:teacher, api_id: "123e4567-e89b-12d3-a456-426614174000") }
-      let!(:other_teacher) { FactoryBot.create(:teacher, api_id: "999e4567-e89b-12d3-a456-426614174999") }
-
-      it { is_expected.to contain_exactly(teacher) }
-    end
-
-    context "when it is a partial API participant ID" do
-      let(:query_string) { "4266141740" }
-      let!(:teacher) { FactoryBot.create(:teacher, api_id: "123e4567-e89b-12d3-a456-426614174000") }
-      let!(:other_teacher) { FactoryBot.create(:teacher, api_id: "999e4567-e89b-12d3-a456-426614174999") }
-
-      it { is_expected.to be_empty }
-    end
-
-    context "when it is an exact API ECT training record ID" do
-      let(:query_string) { "523e4567-e89b-12d3-a456-426614174000" }
-      let!(:teacher) { FactoryBot.create(:teacher, api_ect_training_record_id: "523e4567-e89b-12d3-a456-426614174000") }
-      let!(:other_teacher) { FactoryBot.create(:teacher, api_ect_training_record_id: "999e4567-e89b-12d3-a456-426614174999") }
-
-      it { is_expected.to contain_exactly(teacher) }
-    end
-
-    context "when it contains a full API ECT training record ID with extra text" do
-      let(:query_string) { "something: 523e4567-e89b-12d3-a456-426614174000" }
-      let!(:teacher) { FactoryBot.create(:teacher, api_ect_training_record_id: "523e4567-e89b-12d3-a456-426614174000") }
-      let!(:other_teacher) { FactoryBot.create(:teacher, api_ect_training_record_id: "999e4567-e89b-12d3-a456-426614174999") }
-
-      it { is_expected.to contain_exactly(teacher) }
-    end
-
-    context "when it is a partial API ECT training record ID" do
-      let(:query_string) { "5766141740" }
-      let!(:teacher) { FactoryBot.create(:teacher, api_ect_training_record_id: "123e4567-e89b-12d3-a456-576614174000") }
-      let!(:other_teacher) { FactoryBot.create(:teacher, api_ect_training_record_id: "999e4567-e89b-12d3-a456-426614174999") }
-
-      it { is_expected.to be_empty }
-    end
-
-    context "when it is an exact API mentor training record ID" do
-      let(:query_string) { "823e4567-e89b-12d3-a456-426614174000" }
-      let!(:teacher) { FactoryBot.create(:teacher, api_mentor_training_record_id: "823e4567-e89b-12d3-a456-426614174000") }
-      let!(:other_teacher) { FactoryBot.create(:teacher, api_mentor_training_record_id: "999e4567-e89b-12d3-a456-426614174999") }
-
-      it { is_expected.to contain_exactly(teacher) }
-    end
-
-    context "when it contains a full API mentor training record ID with extra text" do
-      let(:query_string) { "id:823e4567-e89b-12d3-a456-426614174000" }
-      let!(:teacher) { FactoryBot.create(:teacher, api_mentor_training_record_id: "823e4567-e89b-12d3-a456-426614174000") }
-      let!(:other_teacher) { FactoryBot.create(:teacher, api_mentor_training_record_id: "999e4567-e89b-12d3-a456-426614174999") }
-
-      it { is_expected.to contain_exactly(teacher) }
-    end
-
-    context "when it is a partial API mentor training record ID" do
-      let(:query_string) { "5766141740" }
-      let!(:teacher) { FactoryBot.create(:teacher, api_mentor_training_record_id: "123e4567-e89b-12d3-a456-576614174000") }
-      let!(:other_teacher) { FactoryBot.create(:teacher, api_mentor_training_record_id: "999e4567-e89b-12d3-a456-426614174999") }
-
-      it { is_expected.to be_empty }
-    end
-
-    context "when it contains multiple API IDs" do
-      let(:query_string) { "123e4567-e89b-12d3-a456-576614174000 something a9285449-7b0e-47f4-b054-2d09c24c7de5 999e4567-e89b-12d3-a456-426614174999" }
-      let!(:teacher) { FactoryBot.create(:teacher, api_id: "123e4567-e89b-12d3-a456-576614174000") }
-      let!(:other_teacher) { FactoryBot.create(:teacher, api_ect_training_record_id: "999e4567-e89b-12d3-a456-426614174999") }
-      let!(:another_teacher) { FactoryBot.create(:teacher, api_mentor_training_record_id: "a9285449-7b0e-47f4-b054-2d09c24c7de5") }
-      let!(:missing_teacher) { FactoryBot.create(:teacher) }
-
-      it { is_expected.to contain_exactly(teacher, other_teacher, another_teacher) }
-    end
-
-    context "when the query is a plain name" do
+    context "when searching by name" do
       let(:query_string) { "Naruto" }
-      let!(:teacher) { FactoryBot.create(:teacher, trs_first_name: "Naruto", trs_last_name: "Uzumaki", api_id: "123e4567-e89b-12d3-a456-426614174000") }
-      let!(:other_teacher) { FactoryBot.create(:teacher, trs_first_name: "Sasuke", trs_last_name: "Uchiha", api_id: "999e4567-e89b-12d3-a456-426614174999") }
+      let!(:teacher) { FactoryBot.create(:teacher, trs_first_name: "Naruto", trs_last_name: "Uzumaki") }
+      let!(:other_teacher) { FactoryBot.create(:teacher, trs_first_name: "Sasuke", trs_last_name: "Uchiha") }
 
       it { is_expected.to contain_exactly(teacher) }
-    end
-
-    context "when the query only contains tsquery punctuation" do
-      let(:query_string) { "<?'" }
-
-      it { is_expected.to be_empty }
     end
 
     context "when filtering ECT rows by contract period across the whole dataset" do
